pyproject.toml
[tool.poetry]
name = "web-modul-02-06"
version = "0.1.0"
description = ""
authors = ["Daryna Mirzaholi <dmirzagholi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
psycopg2-binary = "^2.9.9"
faker = "^23.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
